Move claim, general and specific also to shared, so we can use the higher namespace for different grammars for arguments based on those.
