+++ grammar/shared/specific.py
@@ -1,6 +1,6 @@
 from grammar.shared import instance, category, prototype, verb
+from grammar.shared.claim import Claim
 from parser import Rule, RuleRef, Literal
-from grammar.claim import Claim
 
 
 class SpecificClaim(Claim):
